Added an application that uses the neighbourhood stats layer
The new layer does not support filtering, but does provide a "properties.stats" field on the information for each element, which is a JSON string containing statistics about all crimes.

Also, tidied up the app.js file, removed some of the commented lines and reindented everything. Also, for this application, removed the CQL filtering based incident filtering for individual neighbourhoods.