Added another application
Functionally identical to the neighbourhoodstats app, but uses vector layer for the neighbourhoods, which makes certain new features possible (for example, front end knows about the exact positioning of the features, how many are on screen and so on).

The vector layer also seems to be faster to load, will investigate using vectors for base layer as well.