Refactoring police distribution flow
Including the new style option for dimming logic. This will also be used later in the normal interaction mode.

The state machine for the application is simple, either selection or interaction. In the first, the user should be able to specify a selection amongst neighbourhood features. In the latter, it is all about interacting with the features.